+++ CHANGELOG.md
@@ -35,6 +35,7 @@ n/a
 * Remove deprecated recycler view adapter framework.
 * `MainActivity` displays a warning at the top of the page if a newer version of the app is available for download. This is done to immediately inform the user about a newer version.
 * Fixed bug where entries were sometimes not displayed after being loaded.
+* Added search functionality that searches entries and details.
 
 <br/>